Refactor project structure to src layout
