+++ environment.yml
@@ -1,0 +1,19 @@
+name: mlccd_viz
+channels:
+  - conda-forge
+  - https://conda.anaconda.org/conda-forge
+dependencies:
+  - astropy
+  - matplotlib
+  - numpy<2
+  - pyside6
+  - superqt
+  - coverage
+  - pytest
+  - opencv
+  - flake8
+  - scikit-image
+  - pip
+  - pip:
+    - numpy<2
+    - git+https://gitlab.com/lbl-anp/mlccd/mlccd_models.git
