+++ experiments/experiments-conda-env.yml
@@ -1,15 +1,1 @@
-name: mlccd_viz
-channels:
-  - conda-forge
-  - https://conda.anaconda.org/conda-forge
-dependencies:
-  - astropy
-  - matplotlib
-  - numpy
-  - pyside6
-  - superqt
-  - coverage
-  - pytest
-  - opencv
-  - flake8
-  - scikit-image
+../environment.yml
